pyproject.toml
[tool.poetry]
name = "car-api-project"
version = "0.1.0"
description = ""
authors = ["adhammhaydat <adhammohidat123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
Django = "^3.2.9"
djangorestframework = "^3.12.4"
black = "^21.11b1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
